pyproject.toml
[project]
name = "wrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "django-cors-headers>=4.6.0",
    "django>=5.1.5",
    "djangorestframework-simplejwt>=5.4.0",
    "djangorestframework>=3.15.2",
]
